+++ src/Scripts/Meta/Constants.lua
@@ -592,3 +592,5 @@ mod.SkellyStatueConversations = {
 mod.ModdedPlayerScaleMultiplier = 0.9
 
 mod.IncreasedZoomFractionOffset = 0.13
+
+mod.ExitToHadesFadeColour = config.z_FadeToBlackEnteringHades and game.Color.Black or game.Color.White